refactor(TSP) : moved tsp-related files in model package
