+++ application/src/main/java/fr/delivrooom/application/model/tsp/TSP1.java
@@ -1,4 +1,4 @@
-package fr.delivrooom.application.tsp.custom;
+package fr.delivrooom.application.model.tsp;
 
 import fr.delivrooom.application.model.Graphe;
